push

+++ src/main/java/com/project/oneshot/command/ContractVO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -13,13 +14,12 @@ import java.util.List;
 @Data
 @Builder
 public class ContractVO { //계약가격 테이블(DB기준)
-
     private Integer productNo; //상품번호
     private Integer employeeNo; //사원번호
     private Integer clientNo; //고객사번호
     private Date contractSdate; //계약시작일
     private Date contractEdate; //계약종료일
-    private Integer contractPrice; //책정거래가
+    private BigDecimal contractPrice; //책정거래가
 
     private Integer contractPriceNo;
     private String clientName;
@@ -33,6 +33,4 @@ public class ContractVO { //계약가격 테이블(DB기준)
     private String employeePhone;
 
     private List<ContractItemVO> contractItems;
-
-
 }
